Fix concurrency issues with tests to single file  for AB#16592
