+++ Testing/functional/tests/cypress/integration/e2e/user/quickLinks.js
@@ -340,3 +340,304 @@ describe("Add Quick Link Modal", () => {
             .and("not.have.class", "v-chip--selected");
     });
 });
+
+// Organ Donor Quick Link tests
+const organDonorQuickLinkCardSelector =
+    "[data-testid=organ-donor-registration-card]";
+const organDonorAddQuickLinkChipSelector =
+    "[data-testid=organ-donor-registration-filter]";
+
+describe("Organ Donor Quick Link", () => {
+    it("Remove and Add Organ Donor Card Quick Link", () => {
+        cy.configureSettings({
+            services: {
+                enabled: true,
+                services: [
+                    {
+                        name: "organDonorRegistration",
+                        enabled: true,
+                    },
+                ],
+            },
+        });
+
+        cy.login(
+            Cypress.env("keycloak.username"),
+            Cypress.env("keycloak.password"),
+            AuthMethod.KeyCloak,
+            homePath
+        );
+
+        cy.log("Verifying organ donor quick link card is present");
+        cy.get(organDonorQuickLinkCardSelector).should(
+            "be.visible",
+            "be.enabled"
+        );
+
+        cy.log("Removing organ donor quick link");
+        cy.get(organDonorQuickLinkCardSelector).within(() => {
+            cy.get(quickLinkMenuButtonSelector)
+                .should("be.visible", "be.enabled")
+                .click();
+        });
+        cy.get(quickLinkRemoveButtonSelector).should("be.visible").click();
+
+        cy.log("Verifying organ donor quick link no longer exists");
+        cy.get(organDonorQuickLinkCardSelector).should("not.exist");
+
+        cy.log("Adding organ donor quick link");
+        cy.get(addQuickLinkButtonSelector)
+            .should("be.visible", "be.enabled")
+            .click();
+        cy.get(organDonorAddQuickLinkChipSelector).should("exist").click();
+        cy.get(addQuickLinkSubmitButtonSelector)
+            .should("be.visible", "be.enabled")
+            .click();
+
+        cy.log("Verifying add quick link button is disabled");
+        cy.get(addQuickLinkButtonSelector).should(
+            "be.visible",
+            "not.be.enabled"
+        );
+
+        cy.log("Verifying organ donor quick link card is present");
+        cy.get(organDonorQuickLinkCardSelector).should(
+            "be.visible",
+            "be.enabled"
+        );
+    });
+});
+
+describe("Disabling organ donor services", () => {
+    const testOrganDonorNotPresent = (cy) => {
+        cy.login(
+            Cypress.env("keycloak.username"),
+            Cypress.env("keycloak.password"),
+            AuthMethod.KeyCloak,
+            homePath
+        );
+        cy.log("Verifying organ donor quicklink not present");
+        cy.get(organDonorQuickLinkCardSelector).should("not.exist");
+
+        cy.log("Verifying add organ donor chip not present");
+        cy.get(addQuickLinkButtonSelector)
+            .should("be.visible", "be.enabled")
+            .click();
+        cy.get(organDonorAddQuickLinkChipSelector).should("not.exist");
+    };
+
+    it("Should not show quick link and settings when organ donor service is disabled", () => {
+        cy.configureSettings({
+            datasets: [
+                {
+                    name: "healthVisit",
+                    enabled: true,
+                },
+            ],
+            services: {
+                services: [
+                    {
+                        name: "organDonorRegistration",
+                        enabled: true,
+                    },
+                ],
+            },
+        });
+        testOrganDonorNotPresent(cy);
+    });
+
+    it("Should not show quick link and settings when organ donor service is disabled", () => {
+        cy.configureSettings({
+            datasets: [
+                {
+                    name: "healthVisit",
+                    enabled: true,
+                },
+            ],
+            services: {
+                enabled: true,
+            },
+        });
+        testOrganDonorNotPresent(cy);
+    });
+});
+
+// Vaccine Card Quick Link tests
+const vaccineCardQuickLinkCardSelector = "[data-testid=bc-vaccine-card-card]";
+const vaccineCardAddQuickLinkChipSelector =
+    "[data-testid=bc-vaccine-card-filter]";
+
+describe("Vaccine Card Quick Link", () => {
+    it("Remove and Add Vaccine Card Quick Link", () => {
+        cy.configureSettings({});
+
+        cy.login(
+            Cypress.env("keycloak.username"),
+            Cypress.env("keycloak.password"),
+            AuthMethod.KeyCloak,
+            homePath
+        );
+
+        cy.log("Verifying vaccine quick link card is present");
+        cy.get(vaccineCardQuickLinkCardSelector).should(
+            "be.visible",
+            "be.enabled"
+        );
+
+        cy.log("Removing vaccine card quick link");
+        cy.get(vaccineCardQuickLinkCardSelector).within(() => {
+            cy.get(quickLinkMenuButtonSelector)
+                .should("be.visible", "be.enabled")
+                .click();
+        });
+        cy.get(quickLinkRemoveButtonSelector).should("be.visible").click();
+
+        cy.log("Verifying vaccine card quick link no longer exists");
+        cy.get(vaccineCardQuickLinkCardSelector).should("not.exist");
+
+        cy.log("Adding vaccine card quick link");
+        cy.get(addQuickLinkButtonSelector)
+            .should("be.visible")
+            .should("be.enabled")
+            .click();
+        cy.get(vaccineCardAddQuickLinkChipSelector).should("exist").click();
+        cy.get(addQuickLinkSubmitButtonSelector)
+            .should("be.visible")
+            .should("be.enabled")
+            .click();
+
+        cy.log("Verifying add quick link button is disabled");
+        cy.get(addQuickLinkButtonSelector)
+            .should("be.visible")
+            .should("not.be.enabled");
+
+        cy.log("Verifying vaccine quick link card is present");
+        cy.get(vaccineCardQuickLinkCardSelector).should(
+            "be.visible",
+            "be.enabled"
+        );
+    });
+});
+
+// Health Connect Registry Quick Link tests
+const healthConnectQuickLinkCardSelector =
+    "[data-testid=health-connect-registry-card]";
+const healthConnectAddQuickLinkChipSelector =
+    "[data-testid=health-connect-registry-filter]";
+const serviceName = "healthConnectRegistry";
+
+describe("Health Connect Registry Card", () => {
+    it("Remove and add health connect registry card quick link", () => {
+        cy.configureSettings({
+            services: {
+                enabled: true,
+                services: [
+                    {
+                        name: serviceName,
+                        enabled: true,
+                    },
+                ],
+            },
+        });
+
+        cy.login(
+            Cypress.env("keycloak.username"),
+            Cypress.env("keycloak.password"),
+            AuthMethod.KeyCloak,
+            homePath
+        );
+
+        cy.log("Verifying health connect quick link card is present");
+        cy.get(healthConnectQuickLinkCardSelector).should(
+            "be.visible",
+            "be.enabled"
+        );
+
+        cy.log("Removing health connect quick link");
+        cy.get(healthConnectQuickLinkCardSelector).within(() => {
+            cy.get(quickLinkMenuButtonSelector)
+                .should("be.visible", "be.enabled")
+                .click();
+        });
+        cy.get(quickLinkRemoveButtonSelector).should("be.visible").click();
+
+        cy.log("Verifying health connect quick link no longer exists");
+        cy.get(healthConnectQuickLinkCardSelector).should("not.exist");
+
+        cy.log("Adding health connect quick link");
+        cy.get(addQuickLinkButtonSelector)
+            .should("be.visible", "be.enabled")
+            .click();
+        cy.get(healthConnectAddQuickLinkChipSelector).should("exist").click();
+        cy.get(addQuickLinkSubmitButtonSelector)
+            .should("be.visible")
+            .should("be.enabled")
+            .click();
+
+        cy.log("Verifying add quick link button is disabled");
+        cy.get(addQuickLinkButtonSelector).should(
+            "be.visible",
+            "not.be.enabled"
+        );
+
+        cy.log("Verifying health connect quick link card is present");
+        cy.get(healthConnectQuickLinkCardSelector).should(
+            "be.visible",
+            "be.enabled"
+        );
+    });
+});
+
+describe("Disabling health connect services", () => {
+    const testQuickLinkNotPresent = (cy) => {
+        cy.login(
+            Cypress.env("keycloak.username"),
+            Cypress.env("keycloak.password"),
+            AuthMethod.KeyCloak,
+            homePath
+        );
+        cy.log("Verifying health connect quicklink not present");
+        cy.get(healthConnectQuickLinkCardSelector).should("not.exist");
+
+        cy.log("Verifying add health connect chip not present");
+        cy.get(addQuickLinkButtonSelector)
+            .should("be.visible", "be.enabled")
+            .click();
+        cy.get(healthConnectAddQuickLinkChipSelector).should("not.exist");
+    };
+
+    it("Should not show quick link and settings when health connect service is disabled", () => {
+        cy.configureSettings({
+            datasets: [
+                {
+                    name: "healthVisit",
+                    enabled: true,
+                },
+            ],
+            services: {
+                services: [
+                    {
+                        name: serviceName,
+                        enabled: true,
+                    },
+                ],
+            },
+        });
+        testQuickLinkNotPresent(cy);
+    });
+
+    it("Should not show quick link and settings when health connect service is disabled", () => {
+        cy.configureSettings({
+            datasets: [
+                {
+                    name: "healthVisit",
+                    enabled: true,
+                },
+            ],
+            services: {
+                enabled: true,
+            },
+        });
+        testQuickLinkNotPresent(cy);
+    });
+});
